Move simple and counterfactual predictions to the model package
