+++ trustyai/model/__init__.py
@@ -1,11 +1,14 @@
 # pylint: disable = import-error, too-few-public-methods, invalid-name, duplicate-code
 """General model classes"""
-from typing import List
+from typing import List, Optional, Tuple
+import uuid as _uuid
 
+from java.lang import Long
 from java.util.concurrent import CompletableFuture, ForkJoinPool
 from jpype import JImplements, JOverride, _jcustomizer, _jclass
 from org.kie.kogito.explainability.model import (
     CounterfactualPrediction as _CounterfactualPrediction,
+    DataDistribution,
     DataDomain as _DataDomain,
     Feature,
     FeatureFactory as _FeatureFactory,
@@ -21,6 +24,8 @@ from org.kie.kogito.explainability.model import (
 from org.kie.kogito.explainability.local.counterfactual.entities import (
     CounterfactualEntity,
 )
+
+from trustyai.model.domain import feature_domain
 
 CounterfactualPrediction = _CounterfactualPrediction
 DataDomain = _DataDomain
@@ -272,3 +277,45 @@ def feature(name: str, dtype: str, value=None) -> Feature:
     else:
         _feature = FeatureFactory.newObjectFeature(name, value)
     return _feature
+
+
+def simple_prediction(
+    input_features: List[Feature],
+    outputs: List[Output],
+) -> SimplePrediction:
+    """Helper to build SimplePrediction"""
+    return SimplePrediction(PredictionInput(input_features), PredictionOutput(outputs))
+
+
+# pylint: disable=too-many-arguments
+def counterfactual_prediction(
+    input_features: List[Feature],
+    outputs: List[Output],
+    domains: List[Optional[Tuple]],
+    constraints: Optional[List[bool]] = None,
+    data_distribution: Optional[DataDistribution] = None,
+    uuid: Optional[_uuid.UUID] = None,
+    timeout: Optional[float] = None,
+) -> CounterfactualPrediction:
+    """Helper to build CounterfactualPrediction"""
+    if not uuid:
+        uuid = _uuid.uuid4()
+    if timeout:
+        timeout = Long(timeout)
+    if not constraints:
+        constraints = [False] * len(input_features)
+
+    # build the feature domains from the Python tuples
+    java_domains = _jclass.JClass("java.util.Arrays").asList(
+        [feature_domain(domain) for domain in domains]
+    )
+
+    return CounterfactualPrediction(
+        PredictionInput(input_features),
+        PredictionOutput(outputs),
+        PredictionFeatureDomain(java_domains),
+        constraints,
+        data_distribution,
+        uuid,
+        timeout,
+    )
